# src/accessi-module/openapi/accessi.openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/accessi/federated-auth/providers": {
      "get": {
        "description": "Riservato a superutente. Restituisce solo catalogo e stato Accessi; issuer, secret e configurazioni SSO restano nel backend.",
        "operationId": "getFederatedProviders",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FederatedProviderListResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Elenca i provider SSO censiti",
        "tags": [
          "Federated authentication"
        ]
      },
      "post": {
        "description": "Riservato a superutente. La chiave provider deve corrispondere alla configurazione stabile del backend; non salvare segreti, issuer o token.",
        "operationId": "createFederatedProvider",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateFederatedProviderRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FederatedProviderResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Censisce un provider SSO",
        "tags": [
          "Federated authentication"
        ]
      }
    },
    "/api/accessi/federated-auth/providers/{provider}": {
      "patch": {
        "description": "Riservato a superutente. La chiave provider Ã¨ immutabile: disabilitarla blocca nuovi login SSO ma conserva i collegamenti storici.",
        "operationId": "updateFederatedProvider",
        "parameters": [
          {
            "name": "provider",
            "required": true,
            "in": "path",
            "description": "Chiave stabile del provider censito.",
            "schema": {
              "example": "azure-ad-acme-produzione",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateFederatedProviderRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FederatedProviderResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Aggiorna o disabilita un provider SSO",
        "tags": [
          "Federated authentication"
        ]
      },
      "delete": {
        "description": "Riservato a superutente. Consentito solo se nessuna identita e collegata al provider; in alternativa disabilitalo.",
        "operationId": "deleteFederatedProvider",
        "parameters": [
          {
            "name": "provider",
            "required": true,
            "in": "path",
            "description": "Chiave stabile del provider censito.",
            "schema": {
              "example": "azure-ad-acme-produzione",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionResponse"
                }
              }
            }
          },
          "404": {
            "description": "Provider non registrato.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Provider ancora utilizzato da identita collegate.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Elimina un provider SSO",
        "tags": [
          "Federated authentication"
        ]
      }
    },
    "/api/accessi/federated-auth/users/{codiceUtente}/identities": {
      "get": {
        "description": "Riservato a superutente. Non restituisce token esterni o claim del provider.",
        "operationId": "getFederatedIdentities",
        "parameters": [
          {
            "name": "codiceUtente",
            "required": true,
            "in": "path",
            "schema": {
              "example": 123,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FederatedIdentityListResponse"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Elenca le identita SSO di un utente",
        "tags": [
          "Federated authentication"
        ]
      },
      "post": {
        "description": "Il master deve fornire un provider e subject ottenuti da una verifica SSO lato backend; questo endpoint non verifica token esterni.",
        "operationId": "linkFederatedIdentity",
        "parameters": [
          {
            "name": "codiceUtente",
            "required": true,
            "in": "path",
            "schema": {
              "example": 123,
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateFederatedIdentityRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FederatedIdentityResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Associa un SSO gia verificato a un utente esistente",
        "tags": [
          "Federated authentication"
        ]
      }
    },
    "/api/accessi/federated-auth/users": {
      "post": {
        "description": "Riservato a superutente. L utente nasce confermato, con identita collegata e senza password salvo opt-in esplicito.",
        "operationId": "createFederatedUser",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateFederatedUserRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FederatedIdentityResponse"
                }
              }
            }
          },
          "400": {
            "description": "Provider o subject non validi.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Provider SSO non censito.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Email o identitÃ  SSO giÃ  associate a un utente.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Errore inatteso durante il provisioning; la risposta contiene FEDERATED_USER_PROVISIONING_FAILED.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Schema Accessi non aggiornato.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Crea un utente gestito e SSO-only",
        "tags": [
          "Federated authentication"
        ]
      }
    },
    "/api/accessi/federated-auth/users/{codiceUtente}/password-login": {
      "patch": {
        "description": "Con false il login locale restituisce PASSWORD_LOGIN_DISABLED; le identita SSO attive restano utilizzabili.",
        "operationId": "setPasswordLoginPolicy",
        "parameters": [
          {
            "name": "codiceUtente",
            "required": true,
            "in": "path",
            "schema": {
              "example": 123,
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetPasswordLoginPolicyRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Abilita o disabilita il login con password",
        "tags": [
          "Federated authentication"
        ]
      }
    },
    "/api/accessi/federated-auth/identities/{identityKey}": {
      "delete": {
        "description": "Soft delete: l associazione resta tracciata ma non puo piu autenticare.",
        "operationId": "disableFederatedIdentity",
        "parameters": [
          {
            "name": "identityKey",
            "required": true,
            "in": "path",
            "description": "SHA-256 restituito dagli endpoint di elenco o creazione.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Disabilita un identita SSO",
        "tags": [
          "Federated authentication"
        ]
      },
      "patch": {
        "description": "Consente di modificare nota e stato attivo del collegamento. identityKey e una chiave tecnica, non un token SSO.",
        "operationId": "updateFederatedIdentity",
        "parameters": [
          {
            "name": "identityKey",
            "required": true,
            "in": "path",
            "description": "SHA-256 restituito dagli endpoint di elenco o creazione.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateFederatedIdentityRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Aggiorna un collegamento SSO",
        "tags": [
          "Federated authentication"
        ]
      }
    },
    "/api/accessi/federated-auth/users/{codiceUtente}/identities/{identityKey}": {
      "delete": {
        "description": "Riservato a superutente. Rimuove il collegamento provider e subject dallâ€™utente, senza eliminare lâ€™utente Accessi.",
        "operationId": "deleteFederatedIdentityPermanently",
        "parameters": [
          {
            "name": "codiceUtente",
            "required": true,
            "in": "path",
            "schema": {
              "example": 123,
              "type": "number"
            }
          },
          {
            "name": "identityKey",
            "required": true,
            "in": "path",
            "description": "SHA-256 del collegamento da eliminare.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionResponse"
                }
              }
            }
          },
          "404": {
            "description": "Collegamento assente o non appartenente allâ€™utente.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Elimina definitivamente un collegamento SSO",
        "tags": [
          "Federated authentication"
        ]
      }
    },
...
  },
  "info": {
    "title": "Accessi API",
    "description": "",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "ActionResponse": {
        "type": "object",
        "properties": {
          "severity": {
            "type": "string",
            "example": "success"
          },
          "status": {
            "type": "number",
            "example": 200
          },
          "statusCode": {
            "type": "number",
            "example": 0
          },
          "message": {
            "type": "string",
            "example": "Operazione completata con successo."
          }
        },
        "required": [
          "severity",
          "status",
          "statusCode",
          "message"
        ]
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "severity": {
            "type": "string",
            "example": "error"
          },
          "status": {
            "type": "number",
            "example": 400
          },
          "statusCode": {
            "type": "number",
            "example": 2
          },
          "message": {
            "type": "string",
            "example": "An error occurred"
          },
          "code": {
            "type": "string",
            "example": "ACCESSI_DATABASE_SCHEMA_OUTDATED",
            "description": "Codice stabile leggibile dal client. Non contiene dettagli tecnici o dati sensibili."
          },
          "error": {
            "type": "string",
            "example": "ACCESSI_REQUEST_ERROR",
            "description": "Campo legacy che riporta il codice stabile dell errore, senza SQL o stack trace."
          },
          "details": {
            "example": [
              "email: deve essere una stringa."
            ],
            "description": "Dettagli di validazione sicuri, presenti solo per ACCESSI_VALIDATION_ERROR.",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "severity",
          "status",
          "statusCode",
          "message",
          "error"
        ]
      },
...
      "FederatedProviderListResponse": {
        "type": "object",
        "properties": {
          "Status": {
            "$ref": "#/components/schemas/Status"
          },
          "Message": {
            "type": "string",
            "example": "Dati recuperati con successo.",
            "description": "Messaggio informativo restituito dalla libreria quando presente."
          },
          "Result": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FederatedProviderResponseDto"
            }
          }
        },
        "required": [
          "Status",
          "Result"
        ]
      },
      "CreateFederatedProviderRequest": {
        "type": "object",
        "properties": {
          "provider": {
            "type": "string",
            "example": "azure-ad-acme-produzione",
            "description": "Chiave stabile ASCII configurata anche nel backend. Deve distinguere tenant e ambiente quando necessario."
          },
          "description": {
            "type": "string",
            "example": "Microsoft Entra ID Acme - Produzione",
            "description": "Descrizione amministrativa del provider mostrata nella console."
          },
          "note": {
            "type": "string",
            "example": "Validazione OIDC gestita dal backend impianti."
          }
        },
        "required": [
          "provider",
          "description"
        ]
      },
      "FederatedProviderResponse": {
        "type": "object",
        "properties": {
          "Status": {
            "$ref": "#/components/schemas/Status"
          },
          "Message": {
            "type": "string",
            "example": "Dati recuperati con successo.",
            "description": "Messaggio informativo restituito dalla libreria quando presente."
          },
          "Result": {
            "$ref": "#/components/schemas/FederatedProviderResponseDto"
          }
        },
        "required": [
          "Status",
          "Result"
        ]
      },
      "UpdateFederatedProviderRequest": {
        "type": "object",
        "properties": {
          "description": {
            "type": "string",
            "example": "Microsoft Entra ID Acme - Produzione"
          },
          "active": {
            "type": "boolean",
            "example": true,
            "description": "Con false il provider e tutte le sue identita non possono autenticare nuove sessioni SSO."
          },
          "note": {
            "type": "string",
            "example": "Manutenzione programmata."
          }
        }
      },
...
      "FederatedIdentityListResponse": {
        "type": "object",
        "properties": {
          "Status": {
            "$ref": "#/components/schemas/Status"
          },
          "Message": {
            "type": "string",
            "example": "Dati recuperati con successo.",
            "description": "Messaggio informativo restituito dalla libreria quando presente."
          },
          "Result": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FederatedIdentityResponseDto"
            }
          }
        },
        "required": [
          "Status",
          "Result"
        ]
      },
      "CreateFederatedIdentityRequest": {
        "type": "object",
        "properties": {
          "provider": {
            "type": "string",
            "example": "azure-ad-acme-produzione",
            "description": "Chiave stabile di un provider già censito nel catalogo Accessi. Deve corrispondere alla configurazione del backend e distinguere tenant e ambiente quando necessario."
          },
          "subject": {
            "type": "string",
            "example": "a1b2c3d4-...",
            "description": "ID stabile della persona restituito dal provider dopo validazione lato backend, ad esempio oid, sub o NameID persistente. Non usare l email o lo username."
          },
          "note": {
            "type": "string",
            "example": "Association approved by IT."
          }
        },
        "required": [
          "provider",
          "subject"
        ]
      },
      "FederatedIdentityResponse": {
        "type": "object",
        "properties": {
          "Status": {
            "$ref": "#/components/schemas/Status"
          },
          "Message": {
            "type": "string",
            "example": "Dati recuperati con successo.",
            "description": "Messaggio informativo restituito dalla libreria quando presente."
          },
          "Result": {
            "$ref": "#/components/schemas/FederatedIdentityResponseDto"
          }
        },
        "required": [
          "Status",
          "Result"
        ]
      },
      "CreateFederatedUserRequest": {
        "type": "object",
        "properties": {
          "provider": {
            "type": "string",
            "example": "azure-ad-acme-produzione",
            "description": "Chiave stabile di un provider già censito nel catalogo Accessi. Deve corrispondere alla configurazione del backend e distinguere tenant e ambiente quando necessario."
          },
          "subject": {
            "type": "string",
            "example": "a1b2c3d4-...",
            "description": "ID stabile della persona restituito dal provider dopo validazione lato backend, ad esempio oid, sub o NameID persistente. Non usare l email o lo username."
          },
          "note": {
            "type": "string",
            "example": "Association approved by IT."
          },
          "user": {
            "description": "Accessi profile, grants and roles assigned by the master user.",
            "allOf": [
              {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            ]
          },
          "passwordLoginEnabled": {
            "type": "boolean",
            "default": false,
            "description": "Enables local password login for the newly created user. Defaults to false (SSO-only)."
          }
        },
        "required": [
          "provider",
          "subject",
          "user"
        ]
      },
...
    }
  }
}